refactoring

+++ templates/init.sh
@@ -1,10 +1,11 @@
 #!/bin/sh
 set -e
 
-kubeadm init --config /etc/kubernetes/master-configuration.yml
+kubeadm reset --force
+kubeadm init --config /etc/kubernetes/configuration.yml
 
 [ -d $HOME/.kube ] || mkdir -p $HOME/.kube
-ln -s /etc/kubernetes/admin.conf $HOME/.kube/config
+[ -f $HOME/.kube/config ] || ln -s /etc/kubernetes/admin.conf $HOME/.kube/config
 
 hostname=$(hostname -s)
 
@@ -12,3 +13,7 @@ until $(kubectl get node $hostname > /dev/null 2>/dev/null); do
   echo "Waiting for api server to respond..."
   sleep 5
 done
+
+kubectl label --overwrite node $hostname ${node_labels}
+[ -f /etc/kubernetes/kube-router.yml ] && kubectl apply -f /etc/kubernetes/kube-router.yml
+kubectl -n kube-system delete ds kube-proxy
